Refactor and improve HTML, CSS and JS files
This commit includes reformatting of index.html for better readability, as well as refactor of JavaScript code to improve its maintainability and performance. CSS styles are adjusted for a better look and feel of elements, changes include increased margins and adjusted colors. Also updated links to the latest versions of FontAwesome and Konva.js libraries. Modified code structures in app.js for better readability and maintainability.